Large accounts are overflowing the data types used in the register code and are causing crashes. #92072, #98802
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/1.6@7696 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/register/register-common.h
@@ -32,8 +32,8 @@
  */
 typedef struct _VirtualCellLocation VirtualCellLocation;
 struct _VirtualCellLocation {
-  short virt_row;
-  short virt_col;
+  int virt_row;
+  int virt_col;
 };
 
 
@@ -55,8 +55,8 @@ gboolean virt_cell_loc_equal (VirtualCellLocation vcl1,
 typedef struct _VirtualLocation VirtualLocation;
 struct _VirtualLocation {
   VirtualCellLocation vcell_loc;
-  short phys_row_offset;
-  short phys_col_offset;
+  int phys_row_offset;
+  int phys_col_offset;
 };
